add:stimulus rate
