+++ rails/app/components/bean_review/component.html.slim
@@ -9,15 +9,15 @@
         span #{time_ago_in_words(@review.created_at)}前
       .meta.review_meta
         span.mr-05em 酸味
-        span.ui.yellow.rating.read.mr-05em data-max-rating= "5" data-rating= "#{@review.acidity}"
+        span.ui.yellow.rating.disabled.mr-05em data-max-rating= "5" data-rating= "#{@review.acidity}"
         span.mr-05em 苦味
-        span.ui.yellow.rating.read.mr-05em data-max-rating= "5" data-rating= "#{@review.bitter}"
+        span.ui.yellow.rating.disabled.mr-05em data-max-rating= "5" data-rating= "#{@review.bitter}"
         span.mr-05em 甘み
-        span.ui.yellow.rating.read.mr-05em data-max-rating= "5" data-rating= "#{@review.sweet}"
+        span.ui.yellow.rating.disabled.mr-05em data-max-rating= "5" data-rating= "#{@review.sweet}"
         span.mr-05em コク
-        span.ui.yellow.rating.read.mr-05em data-max-rating= "5" data-rating= "#{@review.rich}"
+        span.ui.yellow.rating.disabled.mr-05em data-max-rating= "5" data-rating= "#{@review.rich}"
         span.mr-05em 風味
-        span.ui.yellow.rating.read.mr-05em data-max-rating= "5" data-rating= "#{@review.flavor}"
+        span.ui.yellow.rating.disabled.mr-05em data-max-rating= "5" data-rating= "#{@review.flavor}"
       .description
         =simple_format(h(@review.content))
       .extra
