--- rails/app/javascript/custom/actiontext.js
@@ -1,1 +1,0 @@
-
