+++ rails/app/views/devise/registrations/my_page.html.slim
@@ -54,7 +54,7 @@
                 .content
                   h3.header =review.title
                   .meta.review_meta
-                    span.ui.star.rating.read data-max-rating= "5" data-rating= "#{review.rate}"
+                    span.ui.star.rating.disabled data-max-rating= "5" data-rating= "#{review.rate}"
                     span.date #{time_ago_in_words(review.created_at)}前
                   .description
                     =simple_format(h(review.content))
